lttng: fix curated emit field type + saturating dim3 packing
Two correctness nits caught in PR review.

1) HSA curated tracepoint hsa_amd_memory_async_copy_on_engine declared
   force_copy_on_sdma as uint64_t in both LTTNG_UST_TP_ARGS and the
   lttng_ust_field_integer line. The DSL classifies the parameter as
   bool, which the codegen canonicalizes to uint32_t with
   (uint32_t)(!!(arg)). The emit helper already passes the
   uint32_t-canonicalized value; widening to uint64_t in the event
   schema wasted 4 bytes per event and disagreed with the canonical
   type. Fix both the TP_ARGS lane and the field declaration to
   uint32_t. (PR2's codegen will be updated separately to emit uint32_t
   directly so regenerating the header reproduces this.)

2) rocm_trace_pack_dims3 in HIP rocm_trace_emit.h masked grid/block
   dims to 16 bits with , silently wrapping any
   dimension >= 65536 with no overflow indicator. ROCM_DIM3_PACK in
   rocm_dim3_pack.h already does the right thing (full 32-bit x lane,
   16-bit saturating y lane, 15-bit saturating z lane, bit-63 overflow
   flag). Rewrite rocm_trace_pack_dims3's body to use the same
   saturation + overflow-flag scheme so consumers can decode dim3
   values from either emit path with one routine. The (uint32_t, uint32_t,
   uint32_t) signature is preserved so this header remains
   self-contained (no <hip/hip_runtime.h> dependency required for the
   dim3 type).

+++ projects/rocr-runtime/runtime/hsa-runtime/lttng/rocm_hsa_curated_tp.h
@@ -112,7 +112,7 @@ lttng_ust_field_integer_hex(uint64_t, completion_signal, completion_signal)
 
 LTTNG_UST_TRACEPOINT_EVENT(
     rocm_hsa, hsa_amd_memory_async_copy_on_engine_args,
-    LTTNG_UST_TP_ARGS(uint64_t, dst, uint64_t, dst_agent, uint64_t, src, uint64_t, src_agent, uint64_t, size, uint32_t, num_dep_signals, uint64_t, completion_signal, int32_t, engine_id, uint64_t, force_copy_on_sdma),
+    LTTNG_UST_TP_ARGS(uint64_t, dst, uint64_t, dst_agent, uint64_t, src, uint64_t, src_agent, uint64_t, size, uint32_t, num_dep_signals, uint64_t, completion_signal, int32_t, engine_id, uint32_t, force_copy_on_sdma),
     LTTNG_UST_TP_FIELDS(
 lttng_ust_field_integer_hex(uint64_t, dst, dst)
 lttng_ust_field_integer_hex(uint64_t, dst_agent, dst_agent)
@@ -122,7 +122,7 @@ lttng_ust_field_integer(uint64_t, size, size)
 lttng_ust_field_integer(uint32_t, num_dep_signals, num_dep_signals)
 lttng_ust_field_integer_hex(uint64_t, completion_signal, completion_signal)
 lttng_ust_field_integer(int32_t, engine_id, engine_id)
-lttng_ust_field_integer(uint64_t, force_copy_on_sdma, force_copy_on_sdma)
+lttng_ust_field_integer(uint32_t, force_copy_on_sdma, force_copy_on_sdma)
     )
 )
